cli/requirements.txt
annotated-types==0.7.0
astroid==3.3.11
black==25.9.0
click==8.3.0
coverage==7.10.7
dill==0.4.0
greenlet==3.2.4
iniconfig==2.1.0
isort==6.1.0
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mccabe==0.7.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pylint==3.3.8
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
pytokens==0.1.10
rich==14.1.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.43
sqlmodel==0.0.25
textual==6.2.1
tomlkit==0.13.3
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==1.0.3
